+++ LittleWarGame/Hatchet.cs
@@ -12,7 +12,7 @@ namespace LittleWarGame
         {
             type = Const.Warrior_Type.attacker;
 
-            myStatus = Const.imageList[Const.Warrior.Hatchet];
+            myStatus = Const.imageList[(int)Const.Warrior.Hatchet];
             myRealStatus = myStatus[Const.Part.A];
 
             setBonus(10);
